+++ packages/gov-ui-kit/src/modules/hooks/useBlockExplorer/index.ts
@@ -1,1 +1,1 @@
-export * from './useBlockExplorer';
+export { ChainEntityType, useBlockExplorer, type IGetChainEntityUrlParams } from './useBlockExplorer';